// backend/src/common/enums/tipo-conflicto.enum.ts
export enum TipoConflicto {
  SIN_DOCENTE = "SIN_DOCENTE",
  SIN_AMBIENTE = "SIN_AMBIENTE",
  CRUCE_DOCENTE = "CRUCE_DOCENTE",
  CRUCE_AMBIENTE = "CRUCE_AMBIENTE",
  CRUCE_GRUPO = "CRUCE_GRUPO",
  CARGA_INSUFICIENTE = "CARGA_INSUFICIENTE",
  CARGA_EXCEDIDA = "CARGA_EXCEDIDA",
  SIN_DISPONIBILIDAD = "SIN_DISPONIBILIDAD",
  CAPACIDAD_INSUFICIENTE = "CAPACIDAD_INSUFICIENTE",
  DESCANSO_MINIMO = "DESCANSO_MINIMO",
  FRANJA_INSTITUCIONAL = "FRANJA_INSTITUCIONAL",
  DIA_NO_LABORABLE = "DIA_NO_LABORABLE",
}